Modificando rotina para mover no momento certo

+++ core/routine.py
@@ -27,14 +27,6 @@ def processar_digitalizado_iop():
         qr_code = ManageQrCode(pdf_path)
         decoded_text = qr_code.get_decoded_text()
 
-        try:
-            # shutil.copyfile(path + filter_name,
-            #                 settings.PATH_MOVE_FILES_TO_LOCAL + settings.PATH_IOP + 'digitalizado/' + filter_name)
-
-            shutil.move(path + filter_name,
-                        settings.PATH_MOVE_FILES_TO + settings.PATH_IOP + 'digitalizado/' + filter_name)
-        except FileNotFoundError:
-            pass
 
         if decoded_text is None:
             # Ignora o resto da iteração caso o QR Code não tenha sido identificado ou não exista
@@ -94,17 +86,24 @@ def processar_digitalizado_iop():
                     arquivo_indexado_serializer = ArquivoIndexadoRoutineSerializer(instance=arquivo,
                                                                                    data=arquivo_indexado_dict,
                                                                                    partial=True)
+                    if arquivo_anterior:
+                        os.remove(settings.PATH_MOVE_FILES_TO + arquivo_anterior)
 
-                    os.remove(settings.PATH_MOVE_FILES_TO + arquivo_anterior)
                 else:
-                    arquivo_anterior = None
                     arquivo_indexado_serializer = ArquivoIndexadoRoutineSerializer(data=arquivo_indexado_dict)
 
                 arquivo_indexado_serializer.is_valid(raise_exception=True)
                 arquivo_indexado_serializer.save()
 
-                # if arquivo_anterior:
-                #     os.remove(settings.PATH_MOVE_FILES_TO + arquivo_anterior)
+                try:
+                    # shutil.copyfile(path + filter_name,
+                    #                 settings.PATH_MOVE_FILES_TO_LOCAL + settings.PATH_IOP + 'digitalizado/' + filter_name)
+
+                    shutil.move(path + filter_name,
+                                settings.PATH_MOVE_FILES_TO + settings.PATH_IOP + 'digitalizado/' + filter_name)
+                except FileNotFoundError:
+                    pass
+
 
             except Exception as e:
                 print(repr(e))
@@ -127,13 +126,7 @@ def processar_prontuario_iop():
 
         text = file_handle.getvalue()
         caminho_base = settings.PATH_IOP + 'prontuario/' + path.stem + '.pdf'
-        try:
-            # shutil.copyfile(fh.name,
-            #                 settings.PATH_MOVE_FILES_TO_LOCAL + caminho_base)
 
-            shutil.move(fh.name, settings.PATH_MOVE_FILES_TO + caminho_base)
-        except Exception as e:
-            pass
         text = text.split('Profissional')[0].split('\n')
 
         try:
@@ -186,14 +179,20 @@ def processar_prontuario_iop():
                                                                                data=arquivo_indexado_dict,
                                                                                partial=True)
 
+                    os.remove(settings.PATH_MOVE_FILES_TO + arquivo_anterior)
                 else:
-                    arquivo_anterior = None
                     indexed_file_serializer = ArquivoIndexadoRoutineSerializer(data=arquivo_indexado_dict)
                 indexed_file_serializer.is_valid(raise_exception=True)
                 indexed_file_serializer.save()
+                try:
+                    # shutil.copyfile(fh.name,
+                    #                 settings.PATH_MOVE_FILES_TO_LOCAL + caminho_base)
 
-                # if arquivo_anterior:
-                #     os.remove(settings.PATH_MOVE_FILES_TO + arquivo_anterior)
+                    shutil.move(fh.name, settings.PATH_MOVE_FILES_TO + caminho_base)
+
+                except Exception as e:
+                    pass
+
 
             except Exception as e:
                 repr(e)
